https://public.tableau.com/app/profile/myles.trowell/viz/Airbnbfullproject_17295428547250/Dashboard1
For this project Our client is starting an Airbnb business in the Seattle area and aims to identify the optimal location to purchase a home that they can also live in part-time. The goal is to use tableau to maximize rental income while considering key factors that influence profitability, such as location, home features (e.g., number of bedrooms), and rental demand trends.